Remove arm64 from lacros dep until arm64 is supported
This CL disables arm64 boards tentatively from all lacros variants
since it is not supported yet and still in experiment.
It could be enabled in the future when it becomes feasible and
testable for sure.

Test dep attributes can be combined to filter target boards and
make it better searchable on dashboard as below:
- "dep:lacros" && "dep:lacros_stable"
- "dep:lacros" && "dep:lacros_unstable"

BUG=chromium:1274891
TEST=tast run <arm64> '("dep:lacros")' => skipped on arm64
TEST=tast run <arm64> '("dep:lacros_unstable")' => skipped on arm64
TEST=tast run <arm64> '("dep:lacros" && "dep:lacros_unstable")' => skipped on arm64

Change-Id: I2ba0a9f0f9d2a305bcdf2b461ceb1037bdbca325
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast/+/3323861
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>
Reviewed-by: Hyung Tae Kim <hyungtaekim@chromium.org>
Commit-Queue: Hyung Tae Kim <hyungtaekim@chromium.org>
Tested-by: Hyung Tae Kim <hyungtaekim@chromium.org>