tast: If the DUT is offline, report test as notrun
If the DUT is offline before running a remote test, log a magic error
string "Test not run" that tast.py will interpret as a NOT_RUN status.
See crrev/c/3350713

Change tast logging to also show as NOTRUN in the test summary.

It will be helpful to be able to tell the difference between a failed
test and a test that couldn't run because of a previous test killed the
DUT.

BUG=b:210136576
TEST=Ran test that powered test off, verified output and results.json

Change-Id: I9a2907f64443a33be7015f38ceda55da8b8c1973
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast/+/3328923
Tested-by: Jeremy Bettis <jbettis@chromium.org>
Auto-Submit: Jeremy Bettis <jbettis@chromium.org>
Reviewed-by: Keigo Oka <oka@chromium.org>