protocol: Add EntityCopyEnd message
The EntityCopyEnd message is to be sent when copy of output files is
completed for an entity. When we make remote bundle call local bundle,
remote bundle will send this message on copy completion, so that tast
CLI can start file copy *after* the copy completion.

BUG=b:187957164
TEST=./fast_build.sh -T

Change-Id: I6fd918499fc3fbd2966fdd2c508e3c2a7c2fc72e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast/+/3364900
Reviewed-by: Seewai Fu <seewaifu@google.com>
Tested-by: Keigo Oka <oka@chromium.org>
Commit-Queue: Keigo Oka <oka@chromium.org>