processor: Update New to take handlers
This is part of the changes to extract minidriver for bundles to use.

For a remote bundle to execute a local bundle, it will assign
handlers different from what we currently use. This patch hence
separates handler creation and assignment.
minidriver's constructor will take handlers factory as a parameter, so
this patch is needed before introduction of minidriver.

BUG=b:187957164
TEST=fast_build.sh -T

Change-Id: Ifa219d0564346c3366b8c6df54c7019920d5e515
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast/+/3364901
Reviewed-by: Seewai Fu <seewaifu@google.com>
Tested-by: Keigo Oka <oka@chromium.org>
Commit-Queue: Keigo Oka <oka@chromium.org>