processor: Add passthrough handler
The passthrough handler copies output files of test bundle, and send
through the receives messages.
The handler will be used by remote bundle to pass through messages from
local bundle to CLI, but stalling copy done message until file copies
is done.

Currently only handler is added, and the EntityCopyEnd message is not
actually sent by the bundle. This will be done is a subsequent patch.

No unit tests are added in this patch, but it will also be added later
as a testing for remote fixture support.

BUG=b:187957164
TEST=./fast_build.sh -T

Change-Id: Idd5773f2b2f5957f9f247e89b8b3a1e5b457d686
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast/+/3364904
Reviewed-by: Seewai Fu <seewaifu@google.com>
Tested-by: Keigo Oka <oka@chromium.org>
Commit-Queue: Keigo Oka <oka@chromium.org>