OWNERS: remove noparent
This file is in the root of the git project, so there are no parents
to inherit.  This file is also being included in other places to get
the list of tast devs, and the noparent setting is being included too
which we don't want.  Remove it here since it isn't needed.

BUG=None
TEST=CQ passes

Change-Id: Iedc37c83c71ba754168a3defc8ca7378e77f08a5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast/+/3345421
Tested-by: Mike Frysinger <vapier@chromium.org>
Auto-Submit: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Shuhei Takahashi <nya@chromium.org>
Commit-Queue: Shuhei Takahashi <nya@chromium.org>