driver: Move creation of local commands to bundleclient
This is part of the changes to extract minidriver for bundles to use.

minidriver should create a client to run tests in a local bundle, but
currently driver owns such a logic, which minidriver cannot use. This
patch addresses the problem, by moving the logic to bundleclient.

BUG=b:187957164
TEST=./fast_build.sh -T

Change-Id: Iccf71478e4791e17baac8d95b42ed6cf396b74d3
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast/+/3365997
Reviewed-by: Seewai Fu <seewaifu@google.com>
Tested-by: Keigo Oka <oka@chromium.org>
Commit-Queue: Keigo Oka <oka@chromium.org>