tast: add Wifi80211ax6E hwdep condition.
In the effort to add 6e tests, we want a wifi6e support precondition
to ensure that the 6e tests can only be executed against 6e devices.

Model is used instead of platform because the platform for Brya devices
is "Generic" according to the chromeos-config for Brya.

BUG=b:204498033
TEST=tast run wifi.AxSimpleConnect.* on non 6e device and observe that
tests are skipped

Change-Id: If771e5853f9698ae1c005d7d5150e408a8aa8631
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast/+/3358978
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>
Reviewed-by: Nicolas Norvez <norvez@chromium.org>
Reviewed-by: Jun Yu <junyuu@chromium.org>
Reviewed-by: Kevin Lund <kglund@google.com>
Tested-by: Billy Zhao <billyzhao@chromium.org>
Commit-Queue: Billy Zhao <billyzhao@chromium.org>