tast: Update UI codelab-3
Update the information and code in the instructions so the codelab is up-to-date.
When the codelab is followed, there is information about the nodes in the instructions
that does not match with the one resulting in ui_tree.txt, causing the test to fail.
The code is also updated so the final full code works and the test created passes.

BUG=None
TEST=Followed codelab creating temp test in wallpaper.Codelab
Ran: tast -verbose run {DUT} wallpaper.Codelab
Expected intermediate steps: ui_tree.txt node information matches the one displayed in
the instructions
Expected final code: test passes

Change-Id: Idd2626b3c9d2828c3d6cd17d247ac29782b3b603
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast/+/3377006
Tested-by: Cristina Guerrero <crisguerrero@chromium.org>
Reviewed-by: Seewai Fu <seewaifu@google.com>
Reviewed-by: Derek Beckett <dbeckett@chromium.org>
Commit-Queue: Cristina Guerrero <crisguerrero@chromium.org>