tast: add Markers to test results log
Add Markers field to the test param and make it available in
the results log for upstreaming it to ResultDB in the next step.

The marker info will allow custom filtering of test results,
and help map the tests with external artifacts. See go/tag-in-tast.

BUG=b:207565340
TEST=FEATURE=test emerge chromeos-base/tast-cmd,
./fast_build.sh -T

Change-Id: Ifc0e7354e8336fa611f8eca95fbfc39f5db33c56
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast/+/3384656
Tested-by: Alvin Jia <alvinjia@google.com>
Reviewed-by: Seewai Fu <seewaifu@google.com>
Commit-Queue: Alvin Jia <alvinjia@google.com>