Move the debugger logic into the command interface.
The debugger was broken by a change which removed the code paths that
were previously using it. Move the debugger into a code path that is
unlikely to change anytime soon, and move more code into the debugger
API so that any future changes are easier to refactor.

BUG=b:217466480
TEST=Ran a local & remote test with the debugger attached and checked
both worked

Change-Id: Ice298f72ba5383bd3070bd80c6825d118991a7a9
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast/+/3430640
Reviewed-by: Keigo Oka <oka@chromium.org>
Tested-by: Matt Stark <msta@google.com>
Commit-Queue: Matt Stark <msta@google.com>