tast-tests/video: skip v4l2 tests on X86 compatible platforms
This CL adds checks to skip v4l2 tests on X86 compatible platforms
(Intel, AMD) with SupportsV4L2StatefulVideoDecoding() and
SupportsV4L2StatelessVideoDecoding() conditions. This is needed because
this test need to be skipped on Intel and AMD platforms.

In addition, this CL adds MT8183 (kukui) as V4L2 stateless API, which
was missed in crrev/c/3448872.

BUG=b:223293743
TEST=tast.video.ChromeStackDecoding.h264_first_mb_in_slice on drawcia, asurada

Change-Id: I46c4e3e0bea92d2f366469125313bd5ff59f54f4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast/+/3525987
Reviewed-by: Miguel Casas-Sanchez <mcasas@chromium.org>
Reviewed-by: Tatsuhisa Yamaguchi <yamaguchi@chromium.org>
Tested-by: Steve Cho <stevecho@chromium.org>
Reviewed-by: Miguel Casas <mcasas@google.com>
Commit-Queue: Steve Cho <stevecho@chromium.org>