tast: Add linting check for wrapping errors in Wrapf()
At present, users can wrap an error inside the errors.Wrapf() message
arg. But, this is not needed as the first arg is the err itself and we
dont need it again. Add a linting check to make sure we can catch this
in presubmits.

BUG=b:187791703
TEST=Added a test and ran fast-build.sh -T to confirm

Change-Id: Id383af18db70ba6c5b50e84a64f6b030d385c430
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast/+/3527346
Reviewed-by: Seewai Fu <seewaifu@google.com>
Tested-by: Srinivas Hegde <srinivashegde@google.com>
Commit-Queue: Srinivas Hegde <srinivashegde@google.com>