crosbundle/hardware: Change fingerprint detection precedure
Fingerprint capability detection based on checking if /dev/cros_fp
exists is flawed because DUT may be in a state in which there is no
communication with the fingerprint MCU eg. test was interrupted and
flashing procedure was not finished.

Actually, the only portable way to detect fingerprint capability is
to check 'cros_config /fingerprint sensor-location' output.

Checking 'cros_config /fingerprint board' is not good option because
there are board variants without fingerprint sensor, but with 'board'
property set.

BUG=b:196304786
TEST=Run fingerprint TAST test eg.
     'tast run <IP> firmware.FpCheckWriteProtect'
     on nocturne, nami, hatch, strongbad

Change-Id: I99f4d91220ba4f3fd313f2002777128341bbf5a7
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast/+/3596079
Reviewed-by: Keigo Oka <oka@chromium.org>
Tested-by: Patryk Duda <patrykd@google.com>
Reviewed-by: Craig Hesling <hesling@chromium.org>
Commit-Queue: Patryk Duda <patrykd@google.com>