Reland "crosbundle/hardware: Change fingerprint detection procedure"
This is a reland of commit 485608038899ad0f1b099f549ca611648b0a8e64

Previously there was a bug on brya-manatee and volteer-manatee images,
so '/dev/cros_fp' was not created. The fingerprint tests were silently
skipped because '/dev/cros_fp' didn't exist. When the original change
was merged, failures started to appear so the patch was reverted
quickly.

Brya-manatee was fixed by cl:3692147 but volteer-manatee is still
broken. Both images use the same kernel, so the problem seems to be
different.

Actually volteer-manatee is a temporary solution to allow ManaTEE
developers test their changes. It's not supposed to be released to
customers and will be eventually deprecated when there are enough brya
devices for developers (see b/229828602#comment23).

Support for fingerprint can't be disabled in cros_config because both
volteer and volteer-manatee uses the same config. The only way is to
skip the tests.

BUG=b:196304786 b:229828602 b:240120426
TEST=Run fingerprint TAST test eg.
     'tast run <IP> firmware.FpCheckWriteProtect'
     on nocturne, nami, hatch, strongbad, zork.
TEST=Run 'tast run <IP> firmware.FpCheckWriteProtect' on volteer-manatee
     and make sure that test is skipped.
TEST=TAST unit tests passed.

Original change's description:
> crosbundle/hardware: Change fingerprint detection precedure
>
> Fingerprint capability detection based on checking if /dev/cros_fp
> exists is flawed because DUT may be in a state in which there is no
> communication with the fingerprint MCU eg. test was interrupted and
> flashing procedure was not finished.
>
> Actually, the only portable way to detect fingerprint capability is
> to check 'cros_config /fingerprint sensor-location' output.
>
> Checking 'cros_config /fingerprint board' is not good option because
> there are board variants without fingerprint sensor, but with 'board'
> property set.
>
> BUG=b:196304786
> TEST=Run fingerprint TAST test eg.
>      'tast run <IP> firmware.FpCheckWriteProtect'
>      on nocturne, nami, hatch, strongbad
>
> Change-Id: I99f4d91220ba4f3fd313f2002777128341bbf5a7
> Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast/+/3596079
> Reviewed-by: Keigo Oka <oka@chromium.org>
> Tested-by: Patryk Duda <patrykd@google.com>
> Reviewed-by: Craig Hesling <hesling@chromium.org>
> Commit-Queue: Patryk Duda <patrykd@google.com>

Change-Id: I5428f60efb34f066b887afbe610f61b2dbb90981
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast/+/3779758
Tested-by: Patryk Duda <patrykd@google.com>
Reviewed-by: Seewai Fu <seewaifu@google.com>
Auto-Submit: Patryk Duda <patrykd@google.com>
Reviewed-by: Craig Hesling <hesling@chromium.org>
Commit-Queue: Craig Hesling <hesling@chromium.org>