Add presubmit to restrict new usages of hardcoded user bind mount
The bind mounts /home/chronos/user and /home/chronos/user/Downloads are
being deprecated and removed. To avoid churn whilst the references are
removed in Tast tests, add a presubmit which disallows the string
literal and exported vars which reference the bind mount whilst removing
their usages.

This also adds a section to the writing tests documentation to provide a
reference for how to retrieve the bind mount in future.

Further context: go/remove-downloads-bind-mount-dd

BUG=b:233132344
TEST=fast_build.sh -T

Change-Id: I0eb4aee6e9198f74af219dadf7ff8d0aa4c73f59
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast/+/3655415
Reviewed-by: Seewai Fu <seewaifu@google.com>
Commit-Queue: Ben Reich <benreich@chromium.org>
Tested-by: Ben Reich <benreich@chromium.org>