add a hwdep condition for HPS
HPS (go/cros-hps) is a sensor peripheral included in certain SKUs of
taeko. It's already described in cros_config and in the HardwareFeatures
proto. Add a hwdep condition for presence of the HPS peripheral, so that
we can skip HPS-related tests on the SKUs without HPS.

BUG=b:214844812
TEST=./fast_build.sh -T

Change-Id: Ia6eb90924c25b336916952eb59c9de40f388df89
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast/+/3716433
Tested-by: Dan Callaghan <dcallagh@chromium.org>
Reviewed-by: Michael Pishchagin <mblsha@google.com>
Commit-Queue: Dan Callaghan <dcallagh@chromium.org>
Reviewed-by: Seewai Fu <seewaifu@google.com>