tast: Add tflite_opencl feature
The elm/hana boards seem to have an issue with OpenCL affecting the
MLBenchmark tests. This is not a priority to fix so we need to skip
them in the meantime. `hwdeps.SkipOnPlatform` is deprecated, and
`hwdeps.SkipOnModel` is known to cause flakes, so we don't want to
use it for a mainline test. This leaves a software feature as the
best way to skip these boards.

BUG=b:233851820
TEST=tast run -buildbundle=crosint <DUT>
     platform.MLBenchmarkMainline.mobilenet_v2_gpu_opencl
TEST=tast unit tests passed.

Cq-Depend: chrome-internal:4844203
Change-Id: I30ad19e1d674824ffbee8da482593715cebe8c94
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast/+/3731628
Reviewed-by: Seewai Fu <seewaifu@google.com>
Reviewed-by: Michael Olney <mpolney@google.com>
Tested-by: Jake Barnes <jakebarnes@chromium.org>
Commit-Queue: Jake Barnes <jakebarnes@chromium.org>