tast: Add a software dependency for arcvm_virtio_blk_data not enabled
This is a part of 3-CL changes (with crrev.com/c/3802512 and
crrev.com/c/3802250) that adds a dependency to
arc.SELinuxHostSideRestorecon.vm so that it can be skipped when
virtio-blk /data is enabled.
This CL adds a software dependency no_virtio_blk_data which requires
the USE flag arcvm_virtio_blk_data not to be used.

BUG=b:233565855
TEST=`./fast_build.sh -T` passed.

Cq-Depend: chromium:3802512
Change-Id: I7c63868fd62d4ffb4645acf15de2eeb2f8b136dc
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast/+/3802665
Commit-Queue: Momoko Hattori <momohatt@chromium.org>
Reviewed-by: Seewai Fu <seewaifu@google.com>
Reviewed-by: Satoshi Niwa <niwa@chromium.org>
Tested-by: Momoko Hattori <momohatt@chromium.org>