tast: Add usb_hid_wake software dependency
HID events from peripherals will wake a suspended ChromeOS device. In
some cases the wake event is registered to a USB device in the linux
device tree, but in other cases the wake event is registered to an
acpi device. This CL adds a software definition which will allow tests
to know if the platform they are using will register the wake event to a
usb or acpi device.

BUG=b:251275943
TEST=tast unit tests passed. Also, running HidWake on octopus results in
"power.HidWake [ SKIP ] missing SoftwareDeps: usb_hid_wake"

Change-Id: I744ccb5f1c0f911762835b750145aeb30da3ccaf
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast/+/3935658
Reviewed-by: Seewai Fu <seewaifu@google.com>
Commit-Queue: Jameson Thies <jthies@google.com>
Tested-by: Jameson Thies <jthies@google.com>