hwdep: Add hwdeps for WiFi VPD SAR tables
The change adds the new hwdeps WifiVpdSar() and WifiNoVpdSar().

These deps utilize a list of all models which make use of VPD SAR
tables. The methodology for determining this list as well as
justification for why this list is stable can be found in comment#10 in
${BUG}.

WifiVpdSar() is satisfied iff the DUT uses VPD SAR tables *and* actually
has the SAR tables within VPD. This is checked by running the command
`vpd -g wifi_sar`.

WiFiNoVpdSar() is satisfied iff the DUT does not use VPD SAR tables at
all.

Notice that a device which is supposed to have VPD SAR tables but
doesn't have them stored in VPD is excluded from both of these
dependencies. This is by design, since this is an invalid state which
triggers erroneous test failures. This state will eventually be
repaired during the DUT repair process, and it should be infrequent
enough that the number of skipped test runs does not interfere with the
signal we get from this test. (Historically, this condition is
encountered during this test ~1 time per month across our whole fleet).

Additional testing for this CL takes place in CL in tast-tests that
actually makes use of these hwdeps.

BUG=b:204199379
TEST=cros_workon --host start cros-config-api && sudo emerge
cros-config-api
TEST=Add temporary log lines and confirm that the vpd data is being
parsed as expected.
TEST=./fast_build.sh -T

Cq-Depend: chromium:3956308
Change-Id: Ida9f7d7f505d44752b867a9fce1595846e18945c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast/+/3956957
Reviewed-by: Nicolas Norvez <norvez@chromium.org>
Reviewed-by: Seewai Fu <seewaifu@google.com>
Commit-Queue: Kevin Lund <kglund@google.com>
Reviewed-by: Billy Zhao <billyzhao@chromium.org>
Tested-by: Kevin Lund <kglund@google.com>