hardware: Remove board check from fingerprint detection
Fingerprint is not supported on volteer-manatee boards. To avoid running
fingerprint tests, CL:3779758 changed fingerprint support detection
procedure to return false if board was volteer-manatee. It was not
possible to change information returned by cros_config, because it was
shared with volteer builds.

Now, volteer-manatee is deprecated, so we can remove this exception now.

BUG=b:240120426
TEST=TAST unit tests passed
TEST=Run any fingerprint test on board that has fingerprint support.
     Make sure that test is run.

Change-Id: Ib08afb0d27f894aaba577f0037901bbcb58d14be
Disallow-Recycled-Builds: test-failures
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast/+/4043123
Reviewed-by: Andrea Grandi <agrandi@google.com>
Reviewed-by: Seewai Fu <seewaifu@google.com>
Tested-by: Patryk Duda <patrykd@google.com>
Commit-Queue: Patryk Duda <patrykd@google.com>