Add an lsb-release access exception for the WiFi test on Reven board.
Reven's WiFi driver is different from the specific driver used for
ChromeOS devices. Therefore, the path of the WiFi module will be
different. In order for the wifi.Driver test to run with different
configurations, it needs to access board information.

To do this, add the WiFi package exception to the allowedPkgs list.

BUG=b:259533503
TEST=tast unit tests passed.

Cq-Depend: chromium:4336721
Change-Id: Iad9b5e51d03003521c799a27ca4e4529d724a0ac
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast/+/4342814
Commit-Queue: Ting-Hao Lin <tinghaolin@google.com>
Reviewed-by: Seewai Fu <seewaifu@google.com>
Tested-by: Ting-Hao Lin <tinghaolin@google.com>
Reviewed-by: Joseph Sussman <josephsussman@google.com>