flex: Add a new wlan device for reven
The known device list in the wlan.go file does not include a wlan
device of reven DUT: satlab-0wgtfqin2033833a-carbon7. As a
result, the wifi.Caps and wifi.Driver tests fail with an unknown
device error.

To resolve this issue, a new device ID should be added to the wlan list.

BUG=b:277266406
TEST=tast unit tests passed.
TEST=tast run <DUT> wifi.Caps.
TEST=tast run <DUT> wifi.Driver.

Change-Id: I57594c2417ba1ea82472f1dd3f90e92e6a3498cf
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast/+/4407002
Reviewed-by: Joseph Sussman <josephsussman@google.com>
Reviewed-by: Seewai Fu <seewaifu@google.com>
Commit-Queue: Joseph Sussman <josephsussman@google.com>
Tested-by: Ting-Hao Lin <tinghaolin@google.com>