software_defs: Add bluetooth_floss
Adds a new software dependency, "bluetooth_floss" that filters out
all boards that do not support floss. The assumption is that if a
board is filtered out it should support floss.

BUG=b:264468943
TEST=tast unit tests pass; verified test skipped when it has this
dependency and a listed board is used; verified test executed when it
has this dependency and the board is not listed

Change-Id: Id0ad8cd15857f3890f5c64eb9d77f1ab5ac69bb4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast/+/4416373
Reviewed-by: Seewai Fu <seewaifu@google.com>
Commit-Queue: Jared Bennett <jaredbennett@google.com>
Tested-by: Jared Bennett <jaredbennett@google.com>
Reviewed-by: Shijin Abraham <shijinabraham@google.com>