tast: Add wifi_router_feature attribute group
Adds a new attribute group, "wifi_router_feature", and sub-attributes
for each feature. This will be used by tauto in wifi test selection
once we've added it to all the wifi tests and the new testbed label
from which we source this data is propagated. The plan is to move this
test filtering to TestbedDeps once that is more easily usable, but
we need this in the interim.

Filtering by router feature like this removes the need for other
labels that filter by router device and provides a more general
solution that aligns with our longer-term goals regarding wificell
test infra. Other, device-specific attributes will be removed once
these are used.

BUG=b:282840889
TEST=tast unit tests pass

Change-Id: Idecceb002c99011e88ddbaa6a2dec4f7b48520cd
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast/+/4712393
Tested-by: Jared Bennett <jaredbennett@google.com>
Reviewed-by: Seewai Fu <seewaifu@google.com>
Commit-Queue: Jared Bennett <jaredbennett@google.com>
Reviewed-by: Billy Zhao <billyzhao@chromium.org>