hwdep: Use variadic parameter in GPU/CPU hwdep APIs
Most of the hwdep APIs are in the variadic form (`...T`), and there's no
reason for calls like CPUSocFamily() to not follow.

This makes the usage more intuitive (from `[]string{"A", "B"}` to `"A",
"B"`), and introduce zero changes elsewhere as type `...T` is equivalent
to type `[]T` from the view within the functions.

BUG=b:294478144
TEST=./fast_build.sh -T

Cq-Depend: chromium:4756232
Change-Id: Iab981621008f990eb5e52b911fa9659e6f793a94
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast/+/4752920
Reviewed-by: Seewai Fu <seewaifu@google.com>
Reviewed-by: Po-Hsien Wang <pwang@chromium.org>
Reviewed-by: Ilja Friedel <ihf@chromium.org>
Tested-by: Fei Shao <fshao@chromium.org>
Commit-Queue: Fei Shao <fshao@chromium.org>