software_deps: Add known_fixed_ssd
This CL adds a "known_fixed_ssd" software def which can be used to
ensure tests that look for the fixed SSD can find it.
On reven we do not set the globs for finding the SSD.
On VM it is either not set, or incorrect as they point to locations
of the real hardware. It is only set for betty, as it is a VM only
overlay.

BUG=b:265023645, b:187787157, b:295118859
TEST=CQ

Change-Id: I2f4f6c38da1d913ec885b478efa080b0a272a813
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast/+/4808510
Reviewed-by: Joseph Sussman <josephsussman@google.com>
Reviewed-by: Li-Yu Yu <aaronyu@google.com>
Commit-Queue: Gwendal Grignou <gwendal@chromium.org>
Tested-by: Gwendal Grignou <gwendal@chromium.org>