hwdep: Add new hwdep HasPDPort(n)
Add a new hwdep HasPDPort(n) that is satisfied if there is a USB-C PD
port with index n.  If port n doesn't exist, or if there is no ChromeEC
and the number of ports can't be determined, the test is skipped.

BUG=b:291571959
TEST=Used hwdep in a test on hayato and sarien

Change-Id: Icce7c06be5dbb2e48020ca3313cda6b1a6cbb6fa
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast/+/4961285
Reviewed-by: Seewai Fu <seewaifu@google.com>
Tested-by: Jeremy Bettis <jbettis@chromium.org>
Auto-Submit: Jeremy Bettis <jbettis@chromium.org>
Commit-Queue: Jeremy Bettis <jbettis@chromium.org>
Reviewed-by: Brett Brotherton <bbrotherton@google.com>