tast: Add software dep for kfence
We need to know if kfence is enabled in the kernel so that we only run
tests for kfence on ones where the config is enabled. We haven't
backported the code beyond v5.10, so make the dep only exist when the
kernel version isn't earlier than that version.

BUG=b:302027202
TEST=fast_build.sh

Change-Id: I909e2e13f3c44d759a2bebd83c3cbd6e9034e627
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast/+/5090089
Reviewed-by: Ian Barkley-Yeung <iby@chromium.org>
Commit-Queue: ChromeOS Auto Retry <chromeos-auto-retry@chromeos-bot.iam.gserviceaccount.com>
Tested-by: Stephen Boyd <swboyd@chromium.org>
Reviewed-by: Zubin Mithra <zsm@chromium.org>
Reviewed-by: Seewai Fu <seewaifu@google.com>