tast: Add a software dependency for the flex_hwis
The tast.platform.FlexHWIS fails on the reven board
because the reven board does not have VPD. This
causes the enrolled fixture to fail when running
the VPD command. The test can pass on the reven-vmtest
because the enrolled fixture does not check the VPD
state on virtual machines. I created a software
dependency that will only allow the test to run on
the reven-vmtest.

BUG=b:308216955
TEST=tast unit test passed

Change-Id: Ie643426dd6ae274ee36e489983f8180c6ee85f38
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast/+/5005122
Reviewed-by: Seewai Fu <seewaifu@google.com>
Reviewed-by: Joseph Sussman <josephsussman@google.com>
Commit-Queue: Ting-Hao Lin <tinghaolin@google.com>
Tested-by: Ting-Hao Lin <tinghaolin@google.com>