tast: Don't spam connection attempts waiting for DUT to reboot
When we reboot a DUT we keep SSHing to the DUT (with a very small
timeout) waiting for the boot ID to change. This works very
reliably. However, if the DUT is taking a while to reboot we will
notice a whole pile of connection events in the logs. Even though
these connection events don't hurt, they gum up the logs a bit. Let's
add a sleep if we connect to the DUT and the boot ID hasn't changed
yet.

This is especially useful with CL:5009122 ("crash: Use new
RebootWithCommand() to crash").

BUG=None
TEST=crash.KernelCrash.mock_consent
TEST=crash.WatchdogCrash

Change-Id: I5898ef834ca2a3ac8a1dc321add4ed0c27cd7506
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast/+/5011275
Reviewed-by: Seewai Fu <seewaifu@google.com>
Tested-by: Douglas Anderson <dianders@chromium.org>
Commit-Queue: Douglas Anderson <dianders@chromium.org>