tast: Make memfd_exec_detection depend on absence of old kernel versions
Invert this software dep so that we don't have to keep updating this for
every new kernel version. Instead, we'll skip on the old chromeos-4.14
kernel, and assume that the kernel code is present otherwise. Once we
delete the 4.14 chromeos kernel we can remove this dep entirely.

BUG=b:266605753
TEST=tast run $chromeos6_6_dut security.MemoryFileExecTelemetry

Change-Id: Ib51cf0706198cd029a7aaf246abd15d10760e09c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast/+/5263245
Reviewed-by: Seewai Fu <seewaifu@google.com>
Commit-Queue: Stephen Boyd <swboyd@chromium.org>
Tested-by: Stephen Boyd <swboyd@chromium.org>
Reviewed-by: Shervin Oloumi <enlightened@chromium.org>