hwdep: Add NoExternalDisplay hwdep
The video team has a number of tests (tast.webrtc.*monitor*) that are
failing on devices in the AV lab which have external monitors. Being
able to skip these tests when there is an both an internal and
external display would be very helpful to reduce the failures. This
adds a hardware dependency for "no external display".

BUG=b:308959885
TEST=tast unit tests passed
TEST=tests skip when an external monitor is attached to the device

Change-Id: Iec2fc969e6c929f78f437246f7dcda8444fab8aa
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast/+/5282764
Commit-Queue: Nathan Hebert <nhebert@chromium.org>
Tested-by: Nathan Hebert <nhebert@chromium.org>
Reviewed-by: Seewai Fu <seewaifu@google.com>
Reviewed-by: Mark PMNKHME <markyacoub@google.com>