tast: Add hwdep option to check for paravirt sched feature
Paravirt scheduling enables guest and host to communicate and share
scheduling information to optimize the scheduling decisions. Initial
implementation is trying to achieve latency improvements on guest vms.
This feature is currently planned to be enabled on arcvm. More details
of the design and implementation here: go/vcpu_boosting_kvm

Add a hwdep option to check for the feature to conditionally enable
tests.

BUG=b:325918094
TEST=build and run the tests

Change-Id: I92911424e73a25108de59765ac4648ec7e75d689
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast/+/5314001
Tested-by: Vineeth Pillai <vineethrp@google.com>
Reviewed-by: Terry Yan <yichiyan@google.com>
Commit-Queue: Vineeth Pillai <vineethrp@google.com>
Reviewed-by: Alex Bergman <abergman@google.com>
Reviewed-by: Judy Hsiao <judyhsiao@google.com>