hwdep: Add support for tiled display
Add support for TiledDisplay hardware feature to allow tests to only
run when a tiled display is connected. This only checks for presence of
a connector with a TILE property, and does not ensure that all tiles of
a tiled display are connected.

BUG=b:331219664
TEST=emerge cros-config-api and fast_build.sh -T

Cq-Depend: chromium:5399139
Change-Id: Ia163c047ad4719372f94bb4f8caf52e356c8e3a4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast/+/5399058
Tested-by: Su Hong Koo <sukoo@chromium.org>
Commit-Queue: Su Hong Koo <sukoo@chromium.org>
Reviewed-by: Mark PMNKHME <markyacoub@google.com>
Reviewed-by: Seewai Fu <seewaifu@google.com>