tast: Make 'shipping_kernel' software dep into a pure use flag
The shipping_kernel software dep doesn't really work because profile
builders re-use the tast-use-flags package from the base profile. In the
case of some board with a 'debug_kernel' profile, the 'debug', 'kcov',
and 'lockdebug' flags won't be set globally because they only apply to
the kernel ebuild. Simplify this logic by making 'shipping_kernel' into
a pure use flag. When the debug profile is built, it will remove the use
flag from the tast-use-flags ebuild so that tests can restrict running
on builds that are using a kernel that's not stuffed with debug features.

BUG=b:334679407
TEST=cros build-packages --board=$BOARD --profile=debug_kernel

Cq-Depend: chromium:5460350
Change-Id: Icc02bcb8e9d0e4d67d6e7d2e10389e33a6c3208d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast/+/5472472
Reviewed-by: Seewai Fu <seewaifu@google.com>
Tested-by: Stephen Boyd <swboyd@chromium.org>
Reviewed-by: Zubin Mithra <zsm@chromium.org>
Reviewed-by: Brian Norris <briannorris@chromium.org>
Commit-Queue: Stephen Boyd <swboyd@chromium.org>