attr: Repurpose group:external-dependency for extended scope
The group:external-dependency used to be for a specific set of the
tests in the past. This CL will repurpose this group for broader scope
of tests that run against external dependencies such as services or
apps.
Having this group will allow us to make it easy to schedule them and
cluster test results.

BUG=b:333401642
TEST=./fast_build.sh -T

Cq-Depend: chromium:5754257
Change-Id: Ia6e1d43b32e8e9743a148066bef3f24ce23b3486
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast/+/5754667
Tested-by: Hyung Tae Kim <hyungtaekim@chromium.org>
Reviewed-by: Terry Yan <yichiyan@google.com>
Reviewed-by: Victor Zhong <xiuwen@google.com>
Commit-Queue: Hyung Tae Kim <hyungtaekim@chromium.org>