Tast HW Dep for background scanning
Currently, all tests that are required to run on devices that support
background scanning are being skipped on unsupported devices based on
the model name. The list of supported model names is maintained in the
tests code, which is messy to maintain as new devices are added.

Create Tast HW Dep for Bluetooth background scanning based on the chipset
so any tests that require devices supporting background scanning can use it.

BUG=b:225966067
TEST=Ran all nearby share e2e tests. tast run —companiondut=cd1:<hostname> <hostname> nearbyshare.*

Change-Id: I2a6fa52cf1be26d7c136f9fa11fd34a04f2fe36a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast/+/5766803
Commit-Queue: Joaquin Marquez <joaquinmarquez@google.com>
Reviewed-by: Michael Hansen <hansenmichael@google.com>
Reviewed-by: Seewai Fu <seewaifu@google.com>
Tested-by: Joaquin Marquez <joaquinmarquez@google.com>