linuxssh: Replace *ssh.Conn with *dut.DUT in RemoteFileDelta
Currently we can use RemoteFileDelta to record current position of a
remote file, and save only the additional file tail to the host. This is
very helpful in saving logs from a remote machine. However, current
implementation stores the SSH connection object in RemoteFileDelta
and uses the same connection to save the file. This will fail if an SSH
reconnection occurs before saving the file.

To solve this problem, we replace the SSH connection object with a DUT
object in RemoteFileDelta, and get the SSH connection dynamically with
Conn() method. This ensures the SSH validity after a reconnection.

BUG=b:352662991
TEST=fast_build.sh -T
TEST=tast run ${DUT} meta.RemoteFileDelta wifi.SimpleConnect.80211a

Cq-Depend: chromium:5768301
Change-Id: I27c8d726b13303273433e79eb9ef0c3fb04462c8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast/+/5767794
Reviewed-by: Terry Yan <yichiyan@google.com>
Tested-by: Zikai Chen <chenzikai@google.com>
Reviewed-by: Jie Jiang <jiejiang@chromium.org>
Commit-Queue: Zikai Chen <chenzikai@google.com>