hwdep: Add a hardware dependency for Vivaldi keyboards
Adds a condition that is only satisfied when a DUT has a keyboard and is
not a board/model that does not support Vivaldi.

The Vivaldi specification originally was published in June 2020. For
boards that have some models that were released before June 2020, each
model was checked manually using crosfleet. Each model that didn't have
a function_row_physmap file in the keyboard's device path (or the
corresponding property in the device tree for ARM devices) was put in
the exclude list. If a model didn't have a keyboard, it was not put
in the exclude list.

More work will need to be done (in a future CL) to ignore all boards
that had all models released before Vivaldi.

BUG=b:357414398
TEST=./fast_build.sh -T

Change-Id: If6ff6e0e7338d45f35ced0e201442b4fcd6938af
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast/+/5794356
Commit-Queue: Joey Holtzman <joeyholtzman@google.com>
Reviewed-by: Henry Barnor <hbarnor@chromium.org>
Reviewed-by: Sean O'Brien <seobrien@chromium.org>
Reviewed-by: David Padlipsky <dpad@google.com>
Tested-by: Joey Holtzman <joeyholtzman@google.com>
Reviewed-by: Seewai Fu <seewaifu@google.com>