linuxssh: Add WaitUntilFileExists
WaitUntilFileExists waits until a file exists using a specified interval
and timeout.

We will use this function to wait for device files after a Servo USB
port is redirected to a DUT. It is being added to Tast directly
because it can be helpful for other testing situations.

BUG=None
TEST=./fast_build.sh -T

Change-Id: I65cc30d2ce007eaa4386d6889982aa7924546606
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast/+/5803670
Reviewed-by: Seewai Fu <seewaifu@google.com>
Commit-Queue: Joey Holtzman <joeyholtzman@google.com>
Tested-by: Joey Holtzman <joeyholtzman@google.com>