tast: change software dependency for arc on reven
Despite enabling the ARC on the reven board, the
Android image installation is initiated in the Chrome
browser (with DLC method) and is subject to hardware
specification restrictions. Consequently, not all reven
devices will have the Android image. Initially, manual
testing is scheduled, with potential adjustments to the
test in the future if necessary.

BUG=b:348453057
TEST=skip the arc tests on reven* board

Change-Id: If3c0dd86b451760fb1da510878df34fd83b548d8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast/+/5889320
Reviewed-by: Muhammad Hasan Khan <mhasank@chromium.org>
Tested-by: TingHao Lin <tinghaolin@google.com>
Commit-Queue: TingHao Lin <tinghaolin@google.com>